# cmd/bnsapi/docs/swagger.yaml
definitions:
...
  handlers.MultipleObjectsResponse:
    properties:
      objects:
        items:
          $ref: '#/definitions/util.KeyValue'
        type: array
    type: object
...
  orm.Model:
    type: object
...
  util.KeyValue:
    properties:
      key:
        $ref: '#/definitions/util.hexbytes'
        type: object
      value:
        $ref: '#/definitions/orm.Model'
        type: object
    type: object
  util.hexbytes:
    items:
      type: integer
    type: array
...
info:
  contact: {}
  license: {}
  title: BNSAPI documentation
paths:
...
  /termdeposit/contracts:
    get:
      description: The term deposit Contract are the contract defining the dates until
        which one can deposit.
      parameters:
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultipleObjectsResponse'
        "404": {}
        "500": {}
      summary: Returns a list of bnsd/x/termdeposit entities.
      tags:
      - IOV token
  /termdeposit/deposits:
    get:
      description: |-
        At most one of the query parameters must exist (excluding offset).
        The query may be filtered by Depositor, in which case it returns all the deposits from the Depositor.
        The query may be filtered by Deposit Contract, in which case it returns all the deposits from this Contract.
        The query may be filtered by Contract ID, in which case it returns the deposits from the Deposit Contract with this ID.
      parameters:
      - description: Depositor address in bech32 (iov1c9eprq0gxdmwl9u25j568zj7ylqgc7ajyu8wxr)
          or hex(C1721181E83376EF978AA4A9A38A5E27C08C7BB2)
        in: query
        name: depositor
        type: string
      - description: Base64 encoded ID
        in: query
        name: contract
        type: string
      - description: Integer encoded Contract ID
        in: query
        name: contract_id
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MultipleObjectsResponse'
        "404": {}
        "500": {}
      summary: Returns a list of bnsd/x/termdeposit Deposit entities (individual deposits).
      tags:
      - IOV token
...
swagger: "2.0"